Feat : Move EmailNotificationConfig.properties email and password to env. variable
