+++ Dockerfile
@@ -19,6 +19,12 @@ ENV FLOPPA_MONGO_CLUSTER_URL=$CLUSTER_URL
 ARG DATABASE
 ENV FLOPPA_MONGO_DATABASE=$DATABASE
 
+ARG HOSTEMAIL
+ENV FLOPPA_HOST_EMAIL=$HOSTEMAIL
+
+ARG HOSTPASSWORD
+ENV FLOPPA_HOST_PASSWORD=$HOSTPASSWORD
+
 COPY --from=build /app/target/floppa-0.0.1-jar-with-dependencies.jar floppa-0.0.1.jar
 
 CMD [ "java", "-jar", "floppa-0.0.1.jar" ]
