+++ docker-compose.yml
@@ -8,6 +8,8 @@ services:
     environment:
       FLOPPA_MONGO_CLUSTER_URL: "mongodb://db:27017/"
       FLOPPA_MONGO_DATABASE: "floppa-dev"
+      FLOPPA_HOST_EMAIL: "floppanotification@gmail.com"
+      FLOPPA_HOST_PASSWORD: "vexwtppdwslsfcra"
 
   db:
     image: mongo:5.0
