+++ src/test/resources/EmailNotificationConfig.properties
@@ -1,7 +1,3 @@
-hostEmail = floppanotification@gmail.com
-
-hostPassword = vexwtppdwslsfcra
-
 smtpDomain = smtp.gmail.com
 
 domainPort = 465
